+++ examples/battleship/main.rs
@@ -1,3 +1,9 @@
+//! Battleship board Game
+//! Usage:
+//! $> ./battleship # No arguments - outputs a TCP port waiting for the opponent to connect.
+//! $> ./battleship <ip:port> #  - connects to a server with the specified ip port
+//! $> ./battleship bot #  play with computer (single player).
+
 // To run, use the command: cargo run --features unstable --example battleship
 #[cfg(all(not(feature = "unstable")))]
 compile_error!("This example relies on unstable features. Run with `--features unstable`");
@@ -6,16 +12,17 @@ mod battlestate;
 mod bot;
 mod channel;
 mod config;
+mod crosshair;
 mod game;
+mod grid;
 mod ship;
 
+use bot::Bot;
+use channel::ChannelType;
+use game::Game;
 use std::{net::TcpListener, thread, time::Duration};
 
-use game::{Game, Player};
-
-use crate::bot::Bot;
-
-fn get_available_tcp_port() -> u16 {
+pub fn get_available_tcp_port() -> u16 {
     for port in 49152..=65535 {
         if TcpListener::bind(&format!("127.0.0.1:{}", port)).is_ok() {
             return port;
@@ -34,7 +41,7 @@ fn main() {
 
     match config {
         "" => {
-            let mut game = Game::init(Player::Server);
+            let mut game = Game::init(ChannelType::Server);
             game.run();
         }
         "bot" => {
@@ -46,12 +53,12 @@ fn main() {
                 let mut bot = Bot::new(port);
                 bot.play();
             });
-            let mut game = Game::init(Player::ServeOnPort(port));
+            let mut game = Game::init(ChannelType::ServeOnPort(port));
             game.run();
             handle.join().unwrap();
         }
         addr => {
-            let mut game = Game::init(Player::Client(addr));
+            let mut game = Game::init(ChannelType::Client(addr));
             game.run();
         }
     }
